+++ resources/project.rb
@@ -1,4 +1,3 @@
-
 #
 # Cookbook:: rundeck
 # Resource:: project