+++ CHANGELOG.md
@@ -3,6 +3,7 @@ Changelog
 
 # Unreleased
 * Complete rewrite of cookbook using custom resources
+* Adds circleci testing
 
 ## 4.2.0
 * Add optional LDAP data bag support