+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Changelog
 # Unreleased
 * Complete rewrite of cookbook using custom resources
 * Adds circleci testing
+* Adds support for Debian 9
 
 ## 4.2.0
 * Add optional LDAP data bag support